ci(*): create `release.yml` (#9)
This pull request introduces a new GitHub Actions workflow for handling
releases and updates the sync-client configuration to include this new
workflow. The most important changes are summarized below:

### Addition of Release Workflow:
*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34R1-R32):
Added a new workflow to automate the release process when a pull request
is closed, provided it meets specific criteria. This includes setting up
environment variables and publishing a release.

### Update to Sync-Client Configuration:
*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1R18-R19):
Updated the sync-client configuration to include the new release
workflow file.