feat(eslint-plugin-mark): create `alt-text` rule (#31)
This pull request introduces a new ESLint rule to enforce the use of
alternative text for images in the `eslint-plugin-mark` package. The
most important changes include adding the new rule, updating the
configuration, and creating corresponding tests and documentation.

### New Rule Addition:
*
[`packages/eslint-plugin-mark/src/rules/alt-text/alt-text.js`](diffhunk://#diff-6560944db191e902d3ac04b45b164f66c5bd84703cfc914c50c75fc942048c9fR1-R96):
Added a new rule to enforce the use of alternative text for images. This
rule checks for `Image`, `ImageReference`, and `Html` nodes to ensure
they contain `alt` text.

### Configuration Updates:
*
[`packages/eslint-plugin-mark/src/configs/all.js`](diffhunk://#diff-dcdf50541cdda584e7b5e20419349e2a8760d86d4d140d12dbd1d97b3084703fR35):
Added the `mark/alt-text` rule to the list of rules enforced by the
`all` configuration.
*
[`packages/eslint-plugin-mark/src/rules/index.js`](diffhunk://#diff-8bd6342bada2eb43bbaea244fd51588f57f5ec17967476e31bbf0408aaa82ed1R1-R6):
Included the new `alt-text` rule in the list of available rules.

### Testing:
*
[`packages/eslint-plugin-mark/src/rules/alt-text/alt-text.test.js`](diffhunk://#diff-4bf5c4182b3bd7fbce9deac2a8aaa8d943d3eff131d6427495172d55c4f55106R1-R151):
Created tests for the `alt-text` rule to ensure it correctly identifies
valid and invalid cases of image elements without alternative text.

### Documentation:
*
[`website/docs/rules/alt-text.md`](diffhunk://#diff-05707a63cf0061d03420be2cfcb93cb508699d7df712e45a57eb6faa2544b93eR1-R105):
Added documentation for the new `alt-text` rule, including examples of
correct and incorrect code, and references to accessibility guidelines.

### Dependency Update:
*
[`packages/eslint-plugin-mark/package.json`](diffhunk://#diff-9b9400561d26b4f60200e53b8fbb54b0ef18955f8348a54f396c6ac386667a0dL79-R80):
Added `cheerio` as a new dependency to parse and manipulate HTML in the
`alt-text` rule.