feat(eslint-plugin-mark): create `heading-id` rule (#36)
This pull request introduces a new ESLint rule to enforce the use of
heading IDs in markdown files and includes various other related
updates. The most important changes include the addition of the new
rule, its integration into the existing configuration, and the creation
of corresponding tests and documentation.

### New Rule Addition:
*
[`packages/eslint-plugin-mark/src/rules/heading-id/heading-id.js`](diffhunk://#diff-64161e7614a45746324c8428c938fb2c11f1cf8d9f91d58b096a51ee7e13fb10R1-R158):
Added a new rule to enforce the use of heading IDs in markdown files.
This rule includes options for customizing delimiters and ignoring
specific heading depths.

### Rule Integration:
*
[`packages/eslint-plugin-mark/src/configs/all.js`](diffhunk://#diff-dcdf50541cdda584e7b5e20419349e2a8760d86d4d140d12dbd1d97b3084703fR37):
Updated the configuration to include the new `mark/heading-id` rule.
*
[`packages/eslint-plugin-mark/src/rules/index.js`](diffhunk://#diff-8bd6342bada2eb43bbaea244fd51588f57f5ec17967476e31bbf0408aaa82ed1R3-R11):
Imported and added the new `heading-id` rule to the list of available
rules.
*
[`packages/eslint-plugin-mark/src/rules/heading-id/index.js`](diffhunk://#diff-de51b055a4bd30f5c73d4396d81b1a308033d72afd776a17a4d6aecd1115b532R1-R3):
Created an index file for the new rule.

### Testing:
*
[`packages/eslint-plugin-mark/src/rules/heading-id/heading-id.test.js`](diffhunk://#diff-41d3fff468d29b876a36603929b8d3fc0d7955966a54462e0d2f941b3d1be571R1-R358):
Added comprehensive tests for the new rule, covering various valid and
invalid cases.

### Documentation:
*
[`website/docs/rules/heading-id.md`](diffhunk://#diff-ed37936a4dbcafb3ae71f55b57e3054eda68340887ef6cb2a566b351f402eabcR1-R148):
Added documentation for the new rule, including usage examples, options,
and prior art references.

### Other Changes:
*
[`.editorconfig`](diffhunk://#diff-0947e2727d6bad8cd0ac4122f5314bb5b04e337393075bc4b5ef143b17fcbd5bR12-R14):
Updated to disable trimming trailing whitespace in markdown files.