feat(eslint-plugin-mark): create `no-irregular-dash` rule (#58)
This pull request introduces a new ESLint rule to disallow irregular
dash characters in Markdown documents and updates the configuration and
documentation accordingly. The changes include adding the new rule to
the existing configurations, implementing the rule logic, and creating
tests and documentation for the new rule.

### New Rule Implementation:

*
[`packages/eslint-plugin-mark/src/rules/no-irregular-dash/no-irregular-dash.js`](diffhunk://#diff-324916246bf592d49c2a91cf38742025a6820671276a9ea0eeb3fc0ac0ac1525R1-R136):
Implemented the new rule to disallow irregular dash characters,
including the rule definition, options, and logic to identify and report
irregular dashes.
*
[`packages/eslint-plugin-mark/src/rules/no-irregular-dash/no-irregular-dash.test.js`](diffhunk://#diff-77b7832c03592830fc4781fa8d296e631479558daf3705058f0662020481f85aR1-R226):
Added tests for the new rule to ensure it correctly identifies valid and
invalid cases, including various options.

### Configuration Updates:

*
[`packages/eslint-plugin-mark/src/configs/all.js`](diffhunk://#diff-dcdf50541cdda584e7b5e20419349e2a8760d86d4d140d12dbd1d97b3084703fR43):
Added the new `mark/no-irregular-dash` rule to the "all" configuration.
*
[`packages/eslint-plugin-mark/src/configs/recommended.js`](diffhunk://#diff-5c2e62dd777b6828656ee264af3302d295f3768516f0cf7dfdbaa939848a0c01R41):
Added the new `mark/no-irregular-dash` rule to the "recommended"
configuration.

### Rule Index Update:

*
[`packages/eslint-plugin-mark/src/rules/index.js`](diffhunk://#diff-8bd6342bada2eb43bbaea244fd51588f57f5ec17967476e31bbf0408aaa82ed1R11):
Imported and exported the new `no-irregular-dash` rule in the rules
index file.
[[1]](diffhunk://#diff-8bd6342bada2eb43bbaea244fd51588f57f5ec17967476e31bbf0408aaa82ed1R11)
[[2]](diffhunk://#diff-8bd6342bada2eb43bbaea244fd51588f57f5ec17967476e31bbf0408aaa82ed1R23)
*
[`packages/eslint-plugin-mark/src/rules/no-irregular-dash/index.js`](diffhunk://#diff-daf9f13ac6e14a9371719ab03a5d215acef98fa8fa7b1ec6bb67c9e9a529badeR1-R3):
Created an index file for the new rule.

### Documentation Updates:

*
[`website/docs/rules/no-irregular-dash.md`](diffhunk://#diff-4e25635aadad8964af9d641d9a705d233c27c4e06066b7fda8bfa3da94738a44R1-R57):
Added documentation for the new `no-irregular-dash` rule, including rule
details, options, and examples.
*
[`website/docs/rules/no-git-conflict-marker.md`](diffhunk://#diff-5e6b1687e0b2cf50aed9dd8bc06805979e4a827b46b4dc3f1c47d33ed2a4e252L72-R72):
Updated the description to clarify the rule's application scope.
*
[`website/docs/rules/no-irregular-whitespace.md`](diffhunk://#diff-1b6362b81ab627fb3d0e02f0c4d30ea6969217d8a171c7ad514695af6dc73536L84-R84):
Updated the description to clarify the rule's application scope.