feat(eslint-plugin-mark): create `en-capitalization` rule (#63)
This pull request introduces a new ESLint rule for enforcing
capitalization at the beginning of sentences in Markdown documents. The
main changes include the implementation of the rule, its configuration,
and associated tests.

### New Rule Implementation:
*
[`packages/eslint-plugin-mark/src/rules/en-capitalization/en-capitalization.js`](diffhunk://#diff-c7bb138240c9515f2c61e36fd888c71d2b70925c4e404862f67776dc125c8e0dR1-R164):
Added a new rule `en-capitalization` to enforce capitalization at the
beginning of sentences. This rule includes options to skip checking
headings and list items.

### Configuration Updates:
*
[`packages/eslint-plugin-mark/src/configs/all.js`](diffhunk://#diff-dcdf50541cdda584e7b5e20419349e2a8760d86d4d140d12dbd1d97b3084703fR37):
Added the `en-capitalization` rule to the `all` configuration.
*
[`packages/eslint-plugin-mark/src/configs/recommended.js`](diffhunk://#diff-5c2e62dd777b6828656ee264af3302d295f3768516f0cf7dfdbaa939848a0c01R37):
Added the `en-capitalization` rule to the `recommended` configuration.
*
[`packages/eslint-plugin-mark/src/rules/index.js`](diffhunk://#diff-8bd6342bada2eb43bbaea244fd51588f57f5ec17967476e31bbf0408aaa82ed1R5):
Imported and registered the `en-capitalization` rule.
[[1]](diffhunk://#diff-8bd6342bada2eb43bbaea244fd51588f57f5ec17967476e31bbf0408aaa82ed1R5)
[[2]](diffhunk://#diff-8bd6342bada2eb43bbaea244fd51588f57f5ec17967476e31bbf0408aaa82ed1R19)

### Testing:
*
[`packages/eslint-plugin-mark/src/rules/en-capitalization/en-capitalization.test.js`](diffhunk://#diff-73f88be16c20332c478e7e006964808e200746cb33038809322941b4a7cd4418R1-R364):
Added tests for the `en-capitalization` rule to ensure it correctly
identifies and fixes violations.

### Documentation:
*
[`website/docs/rules/en-capitalization.md`](diffhunk://#diff-4b6a2acfdfa4061a0753d250714f81f9da08efddd15e978ea84ebefda53d313dR1-R65):
Created documentation for the `en-capitalization` rule, including
examples of correct and incorrect usage, and configuration options.