feat(eslint-plugin-mark): create `allowed-heading` rule (#66)
This pull request introduces a new rule to the `eslint-plugin-mark`
package, specifically for enforcing allowed text in Markdown headings.
The changes include the implementation of the rule, its configuration,
and associated tests.

New Rule Implementation:

*
[`packages/eslint-plugin-mark/src/rules/allowed-heading/allowed-heading.js`](diffhunk://#diff-d5951c4773cae305a950c0c4cc9ccb7ae0edd097899fe6b6a095e19a15c7f34fR1-R177):
Added a new rule to enforce the use of allowed text for headings. This
includes defining the rule's metadata, schema, default options, and the
logic for checking heading text against allowed values.

Configuration Updates:

*
[`packages/eslint-plugin-mark/src/configs/all.js`](diffhunk://#diff-dcdf50541cdda584e7b5e20419349e2a8760d86d4d140d12dbd1d97b3084703fR35):
Added the new `mark/allowed-heading` rule to the configuration with a
warning level.

Integration and Export:

*
[`packages/eslint-plugin-mark/src/rules/allowed-heading/index.js`](diffhunk://#diff-aab247178f553fbf90ac9695433442c4ec48e9d6c2939b6bf64a813f6bb4743eR1-R3):
Exported the new `allowed-heading` rule.
*
[`packages/eslint-plugin-mark/src/rules/index.js`](diffhunk://#diff-8bd6342bada2eb43bbaea244fd51588f57f5ec17967476e31bbf0408aaa82ed1R3):
Imported and added the `allowed-heading` rule to the list of available
rules.
[[1]](diffhunk://#diff-8bd6342bada2eb43bbaea244fd51588f57f5ec17967476e31bbf0408aaa82ed1R3)
[[2]](diffhunk://#diff-8bd6342bada2eb43bbaea244fd51588f57f5ec17967476e31bbf0408aaa82ed1R18)

Documentation:

*
[`website/docs/rules/allowed-heading.md`](diffhunk://#diff-d21baabc6677a5c86bfc5b7bde6c2ad1513319c23af9543680769d0301d3c584R1-R47):
Added documentation for the new `allowed-heading` rule, describing its
purpose, options, and use cases.

Testing:

*
[`packages/eslint-plugin-mark/src/rules/allowed-heading/allowed-heading.test.js`](diffhunk://#diff-108d8803db994a54f59b38537f8dd56bc2ad4c63fa037c016cd0af9772e56d43R1-R358):
Added tests for the `allowed-heading` rule, covering both valid and
invalid cases.