feat(eslint-plugin-mark): create `image-title` rule (#67)
This pull request introduces a new rule to enforce the use of title
attributes for images in Markdown documents. It includes the
implementation of the rule, updates to the configuration files, and
corresponding tests and documentation.

New rule implementation:

*
[`packages/eslint-plugin-mark/src/rules/image-title/image-title.js`](diffhunk://#diff-d8ba748806ee139927f1d01c6a50e08e3e11fbd5e641eb0d9341b176862176faR1-R95):
Added a new rule to enforce the use of title attributes for images. This
rule checks standard Markdown image syntax, image reference definitions,
and HTML image tags.
*
[`packages/eslint-plugin-mark/src/rules/image-title/image-title.test.js`](diffhunk://#diff-993e4680f978eda002e34c282e1e35cd70c4edfd098f7278ae1a2f640fae5152R1-R177):
Added tests for the new image title rule to ensure it correctly
identifies images without title attributes and validates those with
them.
*
[`packages/eslint-plugin-mark/src/rules/image-title/index.js`](diffhunk://#diff-cc195e97e2eeaa58d8fedc77534fbd9518eb265ab3c2c0817339d2e33d968d31R1-R3):
Exported the new image title rule.

Configuration updates:

*
[`packages/eslint-plugin-mark/src/rules/index.js`](diffhunk://#diff-8bd6342bada2eb43bbaea244fd51588f57f5ec17967476e31bbf0408aaa82ed1R8):
Updated to include the new image title rule in the list of available
rules.
[[1]](diffhunk://#diff-8bd6342bada2eb43bbaea244fd51588f57f5ec17967476e31bbf0408aaa82ed1R8)
[[2]](diffhunk://#diff-8bd6342bada2eb43bbaea244fd51588f57f5ec17967476e31bbf0408aaa82ed1R24)
*
[`packages/eslint-plugin-mark/src/configs/all.js`](diffhunk://#diff-dcdf50541cdda584e7b5e20419349e2a8760d86d4d140d12dbd1d97b3084703fR40):
Set the new image title rule to 'error' in the configuration.

Documentation:

*
[`website/docs/rules/image-title.md`](diffhunk://#diff-1200490758bbba4ba6ca0f835674d7b1ff2ff8cf18eec33b99e4d3aa529d5249R1-R62):
Added documentation for the new image title rule, detailing its purpose,
examples of correct and incorrect usage, and scenarios for when not to
use it.