chore(*): update README badge and ESLint config file (#131)
This pull request includes updates to the project's documentation and
configuration files. The most important changes involve improving the
appearance and functionality of badges in the `README.md` file and
simplifying the ESLint configuration by utilizing a utility function.

### Documentation Updates:
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L3-R8):
Updated badge URLs to use a more visually appealing and consistent
style, added a badge for the `next` version of the package, and improved
the readability of the badges by including logos and custom colors.

### Configuration Simplification:
*
[`eslint.config.mjs`](diffhunk://#diff-9601a8f6c734c2001be34a2361f76946d19a39a709b5e8c624a2a5a0aade05f2L1-R7):
Replaced the manual configuration of global ignores with the
`globalIgnores` utility function to streamline the ESLint configuration.