chore(*): remove `bananass-utils-vitepress` dependency (#182)
This pull request updates the Google Analytics integration in the
VitePress configuration and removes the dependency on
`bananass-utils-vitepress`. The most important changes include replacing
the `generateGoogleAnalyticsScript` utility with a custom implementation
and cleaning up related dependencies.

### Google Analytics Integration:

*
[`website/.vitepress/config.js`](diffhunk://#diff-44d6965b3c2f9329fe9aacf68bd5c4f638b1ff0d4b1b08649486193c906d68d1L87-R97):
Replaced the `generateGoogleAnalyticsScript` utility with a custom
implementation for Google Analytics, directly embedding the script tags
in the configuration.

### Dependency Cleanup:

*
[`website/.vitepress/config.js`](diffhunk://#diff-44d6965b3c2f9329fe9aacf68bd5c4f638b1ff0d4b1b08649486193c906d68d1L12):
Removed the import statement for `generateGoogleAnalyticsScript` from
`bananass-utils-vitepress/head`.
*
[`website/package.json`](diffhunk://#diff-fae242fbf77a8a9d52625664bc36ea12316586f8a716b41137d897a2b7e3df76L15):
Removed the `bananass-utils-vitepress` dependency from the project's
package manifest.

### Minor Updates:

*
[`website/.vitepress/theme/index.js`](diffhunk://#diff-ad87dc71875fb22dd87369888cf1745d026de7048d6cd2ebe18c3c16751dbdbeL3):
Removed an unnecessary comment line in the theme entry file.