chore(*): move `FUNDING.yml` to `.github` repository (#192)
This pull request removes references to the `FUNDING.yml` file from the
repository and its synchronization configuration. The changes ensure
that the funding configuration is no longer part of the repository or
its synced files.

### Removal of funding configuration:

*
[`.github/FUNDING.yml`](diffhunk://#diff-07985fdcade0e64d11482724879a644f07879ba61b8fb6c6119e1b1902b72ae4L1-L2):
Deleted the funding configuration, including references to `github` and
`open_collective` funding sources.
*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1L29-L30):
Removed the synchronization entry for `FUNDING.yml`, preventing it from
being synced to the `configs/.github` directory.