chore(deps-dev): bump `vitepress` to `alpha` (#220)
This pull request includes updates to the website's styling and
dependencies. The most important changes involve adding hover effects to
the `DocSearch-Button` and upgrading the `vitepress` dependency to a
newer version.

### Styling enhancements:
*
[`website/.vitepress/theme/style.css`](diffhunk://#diff-6759218b54439d3be31de7fd72687fb762da52e8a4a2d0f818851dd0e8d55f7fR76-R84):
Added hover effects to the `DocSearch-Button` with a border transition
and color change to improve interactivity.

### Dependency updates:
*
[`website/package.json`](diffhunk://#diff-fae242fbf77a8a9d52625664bc36ea12316586f8a716b41137d897a2b7e3df76L18-R18):
Upgraded `vitepress` from version `^1.6.3` to `^2.0.0-alpha.9` to
utilize new features and improvements.