fix(eslint-plugin-mark)!: update ESLint peer dependency version to `^9.15.0` (#293)
This pull request updates the minimum supported version of ESLint for
the `eslint-plugin-mark` package and its documentation. The minimum
required ESLint version is raised from `9.0.0` to `9.15.0`, and all
relevant documentation is updated to reflect this change.

Dependency update:

* Increased the minimum supported ESLint version in `peerDependencies`
from `^9.0.0` to `^9.15.0` in `package.json`.

Documentation updates:

* Updated the configuration documentation to specify support for ESLint
`v9.15.0` and above, including a link to the release.
* Updated the quick info and dependency version documentation to reflect
the new ESLint minimum version.
[[1]](diffhunk://#diff-e96558082390e38b1c5e2e3df94bced778e823b82720535a6a7728a4790caf0eL6-R6)
[[2]](diffhunk://#diff-e96558082390e38b1c5e2e3df94bced778e823b82720535a6a7728a4790caf0eL15-R15)