fix(*): add optional `peerDependenciesMeta` for `eslint` (#396)
This pull request includes minor dependency updates and metadata
improvements to the `package.json` files. The changes focus on keeping
dependencies up-to-date and improving package configuration for peer
dependencies.

Dependency updates:

* Updated the `textlint-rule-allowed-uris` dependency version from
`^2.0.1` to `^2.0.2` in `package.json` to ensure the latest fixes and
features are included.

Package configuration improvements:

* Added the `peerDependenciesMeta` field to
`packages/eslint-plugin-mark/package.json`, marking `eslint` as an
optional peer dependency to improve flexibility for consumers of the
package.