chore(*): replace `textlint` with `eslint-plugin-mark` (#397)
This pull request removes the use of Textlint for Markdown linting and
refines the configuration for Markdown files. The most important changes
are the removal of Textlint-related code, dependencies, and
configuration, and the introduction of an ESLint rule to handle
previously enforced link restrictions.

**Markdown linting and configuration:**

* Removed Textlint integration from `lint-staged.config.mjs` and
`package.json` scripts, so only Markdownlint is used for `.md` files
now.
[[1]](diffhunk://#diff-7de6c7dbbd1b286aac2271abff52aed05b83379c998f2158f357046f9fd06c7dL7-R7)
[[2]](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L35-R35)
* Deleted the `.textlintrc.cjs` configuration file and removed its sync
mapping in `.github/sync-client.yml`.
[[1]](diffhunk://#diff-89069f4de139f1b360a4a5e68319d8ec8fc403760afa416ac1837838991121a8L1-L9)
[[2]](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1L56-L57)
* Removed `textlint` and `textlint-rule-allowed-uris` dependencies from
`package.json`.

**ESLint enhancements for Markdown:**

* Added a global ESLint rule for Markdown files (`mark/allow-link-url`)
to disallow local URLs, replacing the previous Textlint restriction.