docs(website): update documentation (#435)
This pull request improves the documentation by adding helpful links and
clarifying references to configuration options in several Markdown
files. The main focus is on making it easier for users to find detailed
information about configuration options and migration steps.

**Documentation improvements:**

* Added a tip in `migration.md` with a link to the configurations
section for more details on using `@eslint/markdown` together with
`eslint-markdown`.
* Updated references to configuration options (`all`, `recommended`,
`stylistic`, `base`) in `versioning.md` to use markdown links to the
relevant sections in the configurations documentation.
[[1]](diffhunk://#diff-1320cdd0a36ffc0b5bf5052ddf81af8592e7d45c36a5afd0b23d6630322a0611L49-R49)
[[2]](diffhunk://#diff-1320cdd0a36ffc0b5bf5052ddf81af8592e7d45c36a5afd0b23d6630322a0611L64-R66)
* Updated the descriptions for the `recommended` and `stylistic` configs
in the rules index to include links to their respective documentation
sections.